combine packs.go
